requirements.txt
alembic
sqlalchemy
python-dotenv
fastapi
uvicorn
psycopg2
email_validator
uuid
pydantic